Create DeleteProductCommand and handler

+++ CleanArch/Application/Products/Delete/DeleteProductCommandhandler.cs
@@ -1,23 +1,25 @@
 ﻿using Domain.Repositories;
 using MediatR;
 using Domain.Entities;
-using Domain;
 
 namespace Application.Command.Products.Delete;
 public class DeleteProductCommandhandler : IRequestHandler<DeleteProductCommand, long>
 {
     private readonly IProductRepository _repository;
-    private readonly IDomainService _domainService;
-    public DeleteProductCommandhandler(IProductRepository repository , IDomainService domainService)
+    public DeleteProductCommandhandler(IProductRepository repository)
     {
         _repository = repository;
-        _domainService = domainService;
     }
 
     public async Task<long> Handle(DeleteProductCommand request, CancellationToken cancellationToken)
     {
         var Product = await _repository.GetById(request.ProductId);
-        Product.Remove(request.ProductId, _domainService);
+        bool IsProductExist = _repository.IsProductExist(request.ProductId);
+
+        if (IsProductExist is false)
+            throw new Exception();
+            
+        Product.Remove(request.ProductId);
         _repository.Remove(Product);
         await _repository.SaveChanges();
         return Product.Id;
